refactor: naming and add var query
