+++ Source/Presentation/Server/Do-Svyazi.Message.Server.Http/Controllers/ChatController.cs
@@ -1,6 +1,6 @@
 ﻿using Do_Svyazi.Message.Application.CQRS.Chats.Queries;
 using Do_Svyazi.Message.Application.Dto.Chats;
-using Do_Svyazi.Message.Server.Http.Extension;
+using Do_Svyazi.Message.Server.Http.Extensions;
 using MediatR;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
@@ -23,7 +23,8 @@ public class ChatController : ControllerBase
     public async Task<ActionResult<int>> GetCountOfUnreadMessages([FromRoute] Guid chatId)
     {
         var user = HttpContext.GetUserModel();
-        var response = await _mediator.Send(new GetChatUserState.Query(user.Id, chatId));
+        var query = new GetChatUserState.Query(user.Id, chatId);
+        var response = await _mediator.Send(query);
 
         return Ok(response.ChatUserState.UnreadMessageCount);
     }
@@ -33,7 +34,8 @@ public class ChatController : ControllerBase
     public async Task<ActionResult<ChatUserStateDto>> GetChatState([FromRoute] Guid chatId)
     {
         var user = HttpContext.GetUserModel();
-        var response = await _mediator.Send(new GetChatUserState.Query(user.Id, chatId));
+        var query = new GetChatUserState.Query(user.Id, chatId);
+        var response = await _mediator.Send(query);
 
         return Ok(response.ChatUserState);
     }
